DetailsScreen

+++ app/src/main/java/com/example/shipping/domain/module/Product.kt
@@ -9,6 +9,6 @@ data class Product(
     @PrimaryKey(autoGenerate = true) val product_id: Long,
     @ColumnInfo("product_img") var product_img: String,
     @ColumnInfo("product_description") var product_description: String,
-    @ColumnInfo("product_price") var product_price: Double,
+    @ColumnInfo("product_price") var product_price: Long,
     @ColumnInfo("product_count") var product_count: Int
 )
